+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Alt + click on bezier points cycles between tangent symmetry modes (Ctrl+click still works)
     * Changing a bezier point from corner to smooth will add tangents if they are missing
     * The import image dialog now allows importing multiple images at once
+    * Implemented the "Reverse" path action
 * I/O:
     * Added support for SVG text-anchor
     * Image sequence export now supports SVG